# src/internal/presentation/handlers/performance.go
package handlers

import (
	"net/http"

	"github.com/sohosai/ultradonguri-server/internal/domain/entities"
	"github.com/sohosai/ultradonguri-server/internal/domain/repositories"
	"github.com/sohosai/ultradonguri-server/internal/infrastructure/telop/websocket"
	"github.com/sohosai/ultradonguri-server/internal/presentation/model/requests"
	"github.com/sohosai/ultradonguri-server/internal/presentation/model/responses"

	"github.com/gin-gonic/gin"
)

type PerformanceHandler struct {
	SceneManager repositories.SceneManager
	TelopManager repositories.TelopManager
	wsService    *websocket.WebSocketHub
}

// PostPerformanceStart godoc
// @Summary      start performance
// @Description  endpoint for start performance
// @Tags         performance
// @Accept       json
// @Produce      json
// @Param performanceStart body requests.PerformanceRequest true "post performance request"
// @Success      200  {object}  responses.SuccessResponse
// @Failure      400  {object}  responses.ErrorResponse
// @Router       /performance/start [post]
func (h *PerformanceHandler) PostPerformanceStart(c *gin.Context) {
	var perf requests.PerformanceRequest
	results := []responses.Result{}
	if err := c.ShouldBindJSON(&perf); err != nil {
		errRes, status := responses.NewErrorResponseAndHTTPStatus(entities.AppError{Message: err.Error(),
			Kind: entities.InvalidFormat})
		c.JSON(status, errRes)
		return
	}

	perfEntity := perf.ToDomainPerformance()

	// Telopを設定する
	h.TelopManager.SetPerformanceTelop(perfEntity)

	// Viewerへ通知
	resp, err := websocket.TypedWebSocketResponse[websocket.PerformanceStartData]{
		Type: websocket.TypePerformanceStart,
		Data: websocket.ToDataPerfStart(perfEntity),
	}.Encode()
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	h.wsService.PushTelop(resp)
	results = append(results, responses.Result{
		Operation: "telop_change",
		Success:   true,
	})

	// SceneをNormalに設定する
	err = h.SceneManager.SetNormalScene()
	results = append(results, responses.Result{
		Operation: "scene_change",
		Success:   err == nil,
	})

	c.JSON(http.StatusOK, responses.SuccessResponse{Message: "OK", Results: results})
}

// PostPerformanceMusic godoc
// @Summary      start performance music
// @Description  endpoint for start performance music
// @Tags         performance
// @Accept       json
// @Produce      json
// @Param performanceMusic body requests.MusicRequest true "post performance music"
// @Success      200  {object}  responses.SuccessResponse
// @Failure      400  {object}  responses.ErrorResponse
// @Router       /performance/music [post]
func (h *PerformanceHandler) PostPerformanceMusic(c *gin.Context) {
	var music requests.MusicRequest
	results := []responses.Result{}
	if err := c.ShouldBindJSON(&music); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	musicEntity := music.ToDomainMusicPost()

	// Telopを設定する
	h.TelopManager.SetMusicTelop(musicEntity)

	// Viewerへの通知
	telopMessage := h.TelopManager.GetCurrentTelopMessage()
	if telopMessage.IsSome() {
		resp, err := websocket.TypedWebSocketResponse[websocket.PerformanceMusicData]{
			Type: websocket.TypePerformanceMusic,
			Data: websocket.ToDataPerfMusic(musicEntity),
		}.Encode()
		if err != nil {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		h.wsService.PushTelop(resp)
		results = append(results, responses.Result{
			Operation: "telop_change",
			Success:   true,
		})
	}

	// should_be_muteに合わせてミュート切り替えをする
	// force_muteで失敗したときなどはerrが帰ってくるがこれは正常。しかし、すべてのエラーを同じように扱ってしまっているので、現状以下の処理はこの関数の一番最後にないといけない
	// err := h.SceneManager.SetMute(musicEntity.ShouldBeMuted)
	// results = append(results, responses.Result{
	// 	Operation: "mute_change",
	// 	Success:   err == nil,
	// })
	// if err != nil {
	// 	//後でエラーを細かくする
	// 	errRes, status := responses.NewErrorResponseAndHTTPStatus(entities.AppError{Message: err.Error(),
	// 		Kind: entities.InvalidFormat})
	// 	c.JSON(status, errRes)
	// 	return
	// }

	c.JSON(http.StatusOK, responses.SuccessResponse{Message: "OK", Results: results})
}